+++ src/snowflake/snowpark/_internal/analyzer/metadata_utils.py
@@ -31,6 +31,10 @@ class PlanMetadata:
         # If quoted_identifiers is not None, then attributes will be None because we can't infer data types.
         assert not (self.attributes is not None and self.quoted_identifiers is not None)
 
+    @property
+    def has_cached_quoted_identifiers(self) -> bool:
+        return self.attributes is not None or self.quoted_identifiers is not None
+
 
 def infer_quoted_identifiers_from_expressions(
     expressions: List[Expression],
@@ -71,6 +75,11 @@ def infer_metadata(
     Infer metadata from the source plan.
     Returns the metadata including attributes (schema) and quoted identifiers (column names).
     """
+    from snowflake.snowpark._internal.analyzer.binary_plan_node import (
+        Join,
+        LeftAnti,
+        LeftSemi,
+    )
     from snowflake.snowpark._internal.analyzer.select_statement import (
         Selectable,
         SelectStatement,
@@ -111,6 +120,26 @@ def infer_metadata(
                 analyzer,
                 df_aliased_col_name_to_real_col_name,
             )
+        # When source_plan is a Join, we only infer its quoted identifiers
+        # if two plans don't have any common quoted identifier
+        # This is conservative and avoids parsing join condition, but works for Snowpark pandas,
+        # because join DataFrames in Snowpark pandas guarantee that the column names are unique.
+        elif isinstance(source_plan, Join):
+            if (
+                isinstance(source_plan.left, SnowflakePlan)
+                and isinstance(source_plan.right, SnowflakePlan)
+                and source_plan.left._metadata.has_cached_quoted_identifiers
+                and source_plan.right._metadata.has_cached_quoted_identifiers
+            ):
+                quoted_identifiers = (
+                    source_plan.left.quoted_identifiers
+                    if isinstance(source_plan.join_type, (LeftAnti, LeftSemi))
+                    else source_plan.left.quoted_identifiers
+                    + source_plan.right.quoted_identifiers
+                )
+                # if there is common quoted identifier, reset it to None
+                if len(quoted_identifiers) != len(set(quoted_identifiers)):
+                    quoted_identifiers = None
         # If source_plan is a SelectStatement, SQL simplifier is enabled
         elif isinstance(source_plan, SelectStatement):
             # When attributes is cached on source_plan, just use it