+++ src/snowflake/snowpark/_internal/utils.py
@@ -224,6 +224,66 @@ class TempObjectType(Enum):
     CTE = "CTE"
 
 
+# More info about all allowed aliases here:
+# https://docs.snowflake.com/en/sql-reference/functions-date-time#label-supported-date-time-parts
+
+DATETIME_PART_TO_ALIASES = {
+    "year": {"year", "y", "yy", "yyy", "yyyy", "yr", "years", "yrs"},
+    "quarter": {"quarter", "q", "qtr", "qtrs", "quarters"},
+    "month": {"month", "mm", "mon", "mons", "months"},
+    "week": {"week", "w", "wk", "weekofyear", "woy", "wy"},
+    "day": {"day", "d", "dd", "days", "dayofmonth"},
+    "hour": {"hour", "h", "hh", "hr", "hours", "hrs"},
+    "minute": {"minute", "m", "mi", "min", "minutes", "mins"},
+    "second": {"second", "s", "sec", "seconds", "secs"},
+    "millisecond": {"millisecond", "ms", "msec", "milliseconds"},
+    "microsecond": {"microsecond", "us", "usec", "microseconds"},
+    "nanosecond": {
+        "nanosecond",
+        "ns",
+        "nsec",
+        "nanosec",
+        "nsecond",
+        "nanoseconds",
+        "nanosecs",
+        "nseconds",
+    },
+    "dayofweek": {"dayofweek", "weekday", "dow", "dw"},
+    "dayofweekiso": {"dayofweekiso", "weekday_iso", "dow_iso", "dw_iso"},
+    "dayofyear": {"dayofyear", "yearday", "doy", "dy"},
+    "weekiso": {"weekiso", "week_iso", "weekofyeariso", "weekofyear_iso"},
+    "yearofweek": {"yearofweek"},
+    "yearofweekiso": {"yearofweekiso"},
+    "epoch_second": {"epoch_second", "epoch", "epoch_seconds"},
+    "epoch_millisecond": {"epoch_millisecond", "epoch_milliseconds"},
+    "epoch_microsecond": {"epoch_microsecond", "epoch_microseconds"},
+    "epoch_nanosecond": {"epoch_nanosecond", "epoch_nanoseconds"},
+    "timezone_hour": {"timezone_hour", "tzh"},
+    "timezone_minute": {"timezone_minute", "tzm"},
+}
+
+DATETIME_PARTS = set(DATETIME_PART_TO_ALIASES.keys())
+ALIASES_TO_DATETIME_PART = {
+    v: k for k, l in DATETIME_PART_TO_ALIASES.items() for v in l
+}
+DATETIME_ALIASES = set(ALIASES_TO_DATETIME_PART.keys())
+
+
+def unalias_datetime_part(part):
+    lowered_part = part.lower()
+    if lowered_part in DATETIME_ALIASES:
+        return ALIASES_TO_DATETIME_PART[lowered_part]
+    else:
+        raise ValueError(f"{part} is not a recognized date or time part.")
+
+
+def parse_duration_string(duration: str) -> Tuple[int, str]:
+    length, unit = duration.split(" ")
+    length = int(length)
+    unit = unalias_datetime_part(unit)
+    return length, unit
+
+
 def validate_object_name(name: str):
     if not SNOWFLAKE_OBJECT_RE_PATTERN.match(name):
         raise SnowparkClientExceptionMessages.GENERAL_INVALID_OBJECT_NAME(name)
@@ -436,6 +496,14 @@ def parse_positional_args_to_list(*inputs: Any) -> List:
         )
     else:
         return [*inputs]
+
+
+def parse_positional_args_to_list_variadic(*inputs: Any) -> Tuple[List, bool]:
+    """Convert the positional arguments to a list, indicating whether to treat the argument list as a variadic list."""
+    if len(inputs) == 1 and isinstance(inputs[0], (list, tuple, set)):
+        return ([*inputs[0]], False)
+    else:
+        return ([*inputs], True)
 
 
 def _hash_file(
@@ -737,6 +805,155 @@ def warning(name: str, text: str, warning_times: int = 1) -> None:
     if name not in warning_dict:
         warning_dict[name] = WarningHelper(warning_times)
     warning_dict[name].warning(text)
+
+
+# TODO(SNOW-1491199) - This method is not covered by tests until the end of phase 0. Drop the pragma when it is covered.
+def infer_ast_enabled_from_global_sessions(func: Callable) -> bool:  # pragma: no cover
+    session = None
+    try:
+        # Multiple default session attempts:
+        session = snowflake.snowpark.session._get_sandbox_conditional_active_session(
+            None
+        )
+        assert session is not None
+    except (
+        snowflake.snowpark.exceptions.SnowparkSessionException,
+        AssertionError,
+    ):
+        # Use modin session retrieval first, as it supports multiple sessions.
+        # Expect this to fail if modin was not installed, for Python 3.8, ... but that's ok.
+        try:
+            import modin.pandas as pd
+
+            session = pd.session
+        except Exception as e:  # noqa: F841
+            try:
+                # Get from default session.
+                from snowflake.snowpark.context import get_active_session
+
+                session = get_active_session()
+            except Exception as e:  # noqa: F841
+                pass
+    finally:
+        if session is None:
+            logging.debug(
+                f"Could not retrieve default session "
+                f"for function {func.__qualname__}, capturing AST by default."
+            )
+            # session has not been created yet. To not lose information, always encode AST.
+            return True  # noqa: B012
+        else:
+            return session.ast_enabled  # noqa: B012
+
+
+def publicapi(func) -> Callable:
+    """decorator to safeguard public APIs with global feature flags."""
+
+    # TODO(SNOW-1491199) - This method is not covered by tests until the end of phase 0. Drop the pragma when it is covered.
+    @functools.wraps(func)
+    def func_call_wrapper(*args, **kwargs):  # pragma: no cover
+        # warning(func.__qualname__, warning_text)
+
+        # Handle AST encoding, by modifying default behavior.
+        # If a function supports AST encoding, it must have a parameter _emit_ast.
+        # If now _emit_ast is passed as part of kwargs (we do not allow for the positional syntax!)
+        # then we use this value directly. If not, but the function supports _emit_ast,
+        # we override _emit_ast with the session parameter.
+        if "_emit_ast" in func.__code__.co_varnames and "_emit_ast" not in kwargs:
+            # No arguments, or single argument with function.
+            if len(args) == 0 or (len(args) == 1 and isinstance(args[0], Callable)):
+                if func.__name__ in {
+                    "udf",
+                    "udtf",
+                    "udaf",
+                    "pandas_udf",
+                    "pandas_udtf",
+                    "sproc",
+                }:
+                    session = kwargs.get("session")
+                    # Lookup session directly as in implementation of these decorators.
+                    session = snowflake.snowpark.session._get_sandbox_conditional_active_session(
+                        session
+                    )
+                    # If session is None, do nothing (i.e., keep encoding AST).
+                    # This happens when the decorator is called before a session is started.
+                    if session is not None:
+                        kwargs["_emit_ast"] = session.ast_enabled
+                # Function passed fully with kwargs only (i.e., not a method - self will always be passed positionally)
+                elif len(kwargs) != 0:
+                    # Check if one of the kwargs holds a session object. If so, retrieve AST enabled from there.
+                    session_vars = [
+                        var
+                        for var in kwargs.values()
+                        if isinstance(var, snowflake.snowpark.session.Session)
+                    ]
+                    if session_vars:
+                        kwargs["_emit_ast"] = session_vars[0].ast_enabled
+                    else:
+                        kwargs["_emit_ast"] = infer_ast_enabled_from_global_sessions(
+                            func
+                        )
+            elif isinstance(args[0], snowflake.snowpark.dataframe.DataFrame):
+                # special case: __init__ called, self._session is then not initialized yet.
+                if func.__qualname__.endswith(".__init__"):
+                    # Try to find a session argument.
+                    session_args = [
+                        arg
+                        for arg in args
+                        if isinstance(arg, snowflake.snowpark.session.Session)
+                    ]
+                    assert (
+                        len(session_args) != 0
+                    ), f"{func.__qualname__} must have at least one session arg."
+                    kwargs["_emit_ast"] = session_args[0].ast_enabled
+                else:
+                    kwargs["_emit_ast"] = args[0]._session.ast_enabled
+            elif isinstance(
+                args[0], snowflake.snowpark.dataframe_reader.DataFrameReader
+            ):
+                if func.__qualname__.endswith(".__init__"):
+                    assert isinstance(
+                        args[1], snowflake.snowpark.session.Session
+                    ), f"{func.__qualname__} second arg must be session."
+                    kwargs["_emit_ast"] = args[1].ast_enabled
+                else:
+                    kwargs["_emit_ast"] = args[0]._session.ast_enabled
+            elif isinstance(
+                args[0], snowflake.snowpark.dataframe_writer.DataFrameWriter
+            ):
+                if func.__qualname__.endswith(".__init__"):
+                    assert isinstance(
+                        args[1], snowflake.snowpark.DataFrame
+                    ), f"{func.__qualname__} second arg must be dataframe."
+                    kwargs["_emit_ast"] = args[1]._session.ast_enabled
+                else:
+                    kwargs["_emit_ast"] = args[0]._dataframe._session.ast_enabled
+            elif isinstance(
+                args[0],
+                (
+                    snowflake.snowpark.dataframe_stat_functions.DataFrameStatFunctions,
+                    snowflake.snowpark.dataframe_analytics_functions.DataFrameAnalyticsFunctions,
+                    snowflake.snowpark.dataframe_na_functions.DataFrameNaFunctions,
+                ),
+            ):
+                kwargs["_emit_ast"] = args[0]._dataframe._session.ast_enabled
+            elif hasattr(args[0], "_session") and args[0]._session is not None:
+                kwargs["_emit_ast"] = args[0]._session.ast_enabled
+            elif isinstance(args[0], snowflake.snowpark.session.Session):
+                kwargs["_emit_ast"] = args[0].ast_enabled
+            elif isinstance(
+                args[0],
+                snowflake.snowpark.relational_grouped_dataframe.RelationalGroupedDataFrame,
+            ):
+                kwargs["_emit_ast"] = args[0]._df._session.ast_enabled
+            else:
+                kwargs["_emit_ast"] = infer_ast_enabled_from_global_sessions(func)
+
+        # TODO: Could modify internal docstring to display that users should not modify the _emit_ast parameter.
+
+        return func(*args, **kwargs)
+
+    return func_call_wrapper
 
 
 def func_decorator(
@@ -1065,6 +1282,11 @@ def prepare_pivot_arguments(
         )
 
     return df, pc, pivot_values, default_on_null
+
+
+def check_flatten_mode(mode: str) -> None:
+    if not isinstance(mode, str) or mode.upper() not in ["OBJECT", "ARRAY", "BOTH"]:
+        raise ValueError("mode must be one of ('OBJECT', 'ARRAY', 'BOTH')")
 
 
 class MissingModin(MissingOptionalDependency):