+++ src/snowflake/snowpark/mock/_plan.py
@@ -4,6 +4,7 @@
 
 import importlib
 import inspect
+import json
 import math
 import re
 import statistics
@@ -229,6 +230,9 @@ class MockExecutionPlan(LogicalPlan):
             PlanQueryType.POST_ACTIONS: self.post_actions,
         }
 
+    def add_aliases(self, to_add: Dict) -> None:
+        self.expr_to_alias.update(to_add)
+
 
 class MockFileOperation(MockExecutionPlan):
     class Operator(str, Enum):
@@ -390,17 +394,19 @@ def handle_function_expression(
             importlib.import_module("snowflake.snowpark.functions"), func.name
         )
     except AttributeError:
-        # this is missing function in snowpark-python, need support for both live and local test
+        original_func = None
+        # raise error depending on whether users has a provided a patch for function not available in snowpark-python
         analyzer.session._conn.log_not_supported_error(
             external_feature_name=func.name,
             error_message=f"Function {func.name} is not supported in snowpark-python.",
-            raise_error=NotImplementedError,
+            raise_error=NotImplementedError if func is None else None,
         )
 
-    signatures = inspect.signature(original_func)
-    spec = inspect.getfullargspec(original_func)
+    to_mock_func = original_func or func.impl
+    signatures = inspect.signature(to_mock_func)
+    spec = inspect.getfullargspec(to_mock_func)
     to_pass_args = []
-    type_hints = typing.get_type_hints(original_func)
+    type_hints = typing.get_type_hints(to_mock_func)
     for idx, key in enumerate(signatures.parameters):
         type_hint = str(type_hints[key])
         keep_literal = "Column" not in type_hint
@@ -449,9 +455,22 @@ def handle_function_expression(
     try:
         result = func(*to_pass_args, row_number=current_row, input_data=input_data)
     except Exception as err:
+        extra_err_info = (
+            (
+                f"\nA patch is provided for '{func.name}' which is not supported in Snowpark Python."
+                f" Please ensure the implementation follows specifications outlined at:"
+                f" https://docs.snowflake.com/en/sql-reference/functions-all and refer to"
+                f" https://github.com/snowflakedb/snowpark-python/blob/main/src/snowflake/snowpark/mock/_functions.py"
+                f" for patterns on creating a compatible patch function for input and output types."
+            )
+            if not original_func and func
+            else ""
+        )
+
         SnowparkLocalTestingException.raise_from_error(
             err,
-            error_message=f"Error executing mocked function '{func.name}'. See error traceback for detailed information.",
+            error_message=f"Error executing mocked function '{func.name}'."
+            f" See error traceback for detailed information.{extra_err_info}",
         )
 
     return result
@@ -871,9 +890,16 @@ def execute_mock_plan(
                     col_name = f"<local_test_internal_{str(exp.value)}>"
                     by_column_expression.append(child_rf[col_name])
                 else:
-                    by_column_expression.append(
-                        child_rf[plan.session._analyzer.analyze(exp)]
-                    )
+                    column_name = plan.session._analyzer.analyze(exp)
+                    if isinstance(exp, FunctionExpression):
+                        materialized_column = calculate_expression(
+                            exp, child_rf, plan.session._analyzer, expr_to_alias
+                        )
+                        # Only function expressions that are a mapping of existing columns can be aggregated on.
+                        # Any increase or reduction in number of rows is an invalid function expression.
+                        if len(materialized_column) == len(child_rf):
+                            child_rf[column_name] = materialized_column
+                    by_column_expression.append(child_rf[column_name])
         except KeyError as e:
             raise SnowparkLocalTestingException(
                 f"This is not a valid group by expression due to exception {e!r}"
@@ -1202,21 +1228,50 @@ def execute_mock_plan(
             else:
                 matched_rows = target
 
+            # the following function is used to flatten the cell object to string which is hashable
+            # as required by pandas.DataFrame.value_counts/drop_duplicates method
+            def flatten_object_cell_func(cell):
+                if isinstance(cell, (dict, list)):
+                    return json.dumps(cell)
+                return cell
+
             # Calculate multi_join
-            matched_count = intermediate[target.columns].value_counts(dropna=False)[
-                matched_rows.apply(tuple, 1)
-            ]
+
+            # 1. count each row occurrences
+            try:
+                # flatten the object cell (list, dict) to string first
+                flatten_intermediate = intermediate.map(flatten_object_cell_func)
+            except AttributeError:  # for backward compatibility with pandas < 2.1.0
+                flatten_intermediate = intermediate.applymap(flatten_object_cell_func)
+
+            rows_value_counts = flatten_intermediate[target.columns].value_counts(
+                dropna=False
+            )
+            # 2. get the row into tuple serving as the key to index the rows_value_counts DF
+            try:
+                key_index = matched_rows.map(flatten_object_cell_func).apply(tuple, 1)
+            except AttributeError:  # for backward compatibility with pandas < 2.1.0
+                key_index = matched_rows.applymap(flatten_object_cell_func).apply(
+                    tuple, 1
+                )
+            matched_count = rows_value_counts[
+                key_index
+            ]  # 3. get the occurrences of the matched rows
             multi_joins = matched_count.where(lambda x: x > 1).count()
 
             # Select rows that match the condition to be updated
-            rows_to_update = intermediate.drop_duplicates(
+
+            # 1. get the index of the rows to update
+            pd_index = flatten_intermediate.drop_duplicates(
                 subset=matched_rows.columns, keep="first"
-            ).reset_index(  # ERROR_ON_NONDETERMINISTIC_UPDATE is by default False, pick one row to update
-                drop=True
+            ).index
+            # 2. get the rows to update
+            rows_to_update = intermediate.loc[pd_index].reset_index(
+                drop=True  # ERROR_ON_NONDETERMINISTIC_UPDATE is by default False, pick one row to update
             )
             rows_to_update.sf_types = intermediate.sf_types
 
-            # Update rows in place
+            # 3. Update rows in place
             for attr, new_expr in source_plan.assignments.items():
                 column_name = analyzer.analyze(attr, expr_to_alias)
                 target_index = target.loc[rows_to_update[ROW_ID]].index