+++ src/snowflake/snowpark/relational_grouped_dataframe.py
@@ -156,22 +156,27 @@ class RelationalGroupedDataFrame:
         used = set()
         unique = [a for a in aliased_agg if a not in used and (used.add(a) or True)]
         aliased_agg = [_alias(a) for a in unique]
+        # Aliases cannot be used in grouping statements, but the child expresion can
+        unaliased_grouping = [
+            expr.child if isinstance(expr, Alias) else expr
+            for expr in self._grouping_exprs
+        ]
 
         if isinstance(self._group_type, _GroupByType):
             group_plan = Aggregate(
-                self._grouping_exprs,
+                unaliased_grouping,
                 aliased_agg,
                 self._df._select_statement or self._df._plan,
             )
         elif isinstance(self._group_type, _RollupType):
             group_plan = Aggregate(
-                [Rollup(self._grouping_exprs)],
+                [Rollup(unaliased_grouping)],
                 aliased_agg,
                 self._df._select_statement or self._df._plan,
             )
         elif isinstance(self._group_type, _CubeType):
             group_plan = Aggregate(
-                [Cube(self._grouping_exprs)],
+                [Cube(unaliased_grouping)],
                 aliased_agg,
                 self._df._select_statement or self._df._plan,
             )
@@ -179,7 +184,7 @@ class RelationalGroupedDataFrame:
             if len(agg_exprs) != 1:
                 raise SnowparkClientExceptionMessages.DF_PIVOT_ONLY_SUPPORT_ONE_AGG_EXPR()
             group_plan = Pivot(
-                self._grouping_exprs,
+                unaliased_grouping,
                 self._group_type.pivot_col,
                 self._group_type.values,
                 agg_exprs,