+++ tests/conftest.py
@@ -5,6 +5,8 @@
 
 import logging
 import os
+import subprocess
+import sys
 from pathlib import Path
 
 import pytest
@@ -28,7 +30,9 @@ def is_excluded_frontend_file(path):
 def pytest_addoption(parser):
     parser.addoption("--disable_sql_simplifier", action="store_true", default=False)
     parser.addoption("--disable_cte_optimization", action="store_true", default=False)
-    parser.addoption("--multithreading_mode", action="store_true", default=False)
+    parser.addoption(
+        "--disable_multithreading_mode", action="store_true", default=False
+    )
     parser.addoption("--skip_sql_count_check", action="store_true", default=False)
     if not any(
         "--local_testing_mode" in opt.names() for opt in parser._anonymous.options
@@ -90,12 +94,21 @@ def cte_optimization_enabled(pytestconfig):
     return not pytestconfig.getoption("disable_cte_optimization")
 
 
+@pytest.fixture(scope="module", autouse=True)
+def proto_generated():
+    """Generate Protobuf Python files automatically"""
+    try:
+        from snowflake.snowpark._internal.proto.generated import ast_pb2  # noqa: F401
+    except ImportError:
+        subprocess.check_call([sys.executable, "-m", "tox", "-e", "protoc"])
+
+
 MULTITHREADING_TEST_MODE_ENABLED = False
 
 
 @pytest.fixture(scope="session", autouse=True)
 def multithreading_mode_enabled(pytestconfig):
-    enabled = pytestconfig.getoption("multithreading_mode")
+    enabled = not pytestconfig.getoption("disable_multithreading_mode")
     global MULTITHREADING_TEST_MODE_ENABLED
     MULTITHREADING_TEST_MODE_ENABLED = enabled
     return enabled